requirements.txt
python-dotenv
langchain>=1.0.5
langchain-core>=1.0.4
langchain-community>=0.4.1
langchain-classic>=1.0.0
langchain-text-splitters>=0.0.1
langchain-huggingface>=0.0.9
langchain-groq>=0.0.4
langgraph>=0.1.12
sentence-transformers
faiss-cpu
pypdf
pymupdf